+++ site/src/tutorial.rs
@@ -8,8 +8,8 @@ use uiua::{Primitive, SysOp, EXAMPLE_UA};
 use uiua_editor::*;
 
 use crate::{
-    markdown::Markdown, other_tutorial::OtherTutorialParams, title_markdown, Challenge, Hd, Prim,
-    Prims,
+    markdown::Markdown, other_tutorial::OtherTutorialParams, title_markdown, Challenge, Hd, Hd3,
+    Prim, Prims,
 };
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Sequence)]
@@ -1591,6 +1591,11 @@ fn TutorialModules() -> impl IntoView {
         <p>"To enter this arrow, you can put a "<code>"~"</code>" after a binding's normal "<code>"←"</code>" or "<code>"="</code>"."</p>
         <p>"Try formatting the following example to see how this works."</p>
         <Editor example="A = +1\nB ← +2\nC =~ +3\nD ←~ +4"/>
+        <Hd3 id="private-scoped-modules">"Private Scoped Modules"</Hd3>
+        <p>"Scoped modules can be made private with special delimiters "<code>"┌╶╶"</code>" and "<code>"└╶╶"</code>". These format from the normal delimiters or "<code>"---"</code>"s followed by a "<code>"~"</code>"."</p>
+        <Editor example="┌─╴A\n  ┌╶╶B\n    C ← 5\n  └╶╶\n└─╴\nA~B~C"/> // Should fail
+        <Editor example = "# Try formatting!\n---A\n  ---~M\n    F = +1\n  ---\n---\nA~M~F 5"/> // Should fail
+        <p>"The formatter will automatically change the closing delimiter to match its corresponding opening delimiter."</p>
 
         <Hd id="git-modules">"Git Modules"</Hd>
         <p>"Modules can be imported from Git repositories. Instead of a path, use a URL prefixed with "<code>"git:"</code>"."</p>